chore: API & doc polish — clippy clean, rustdoc, README, examples, CHANGELOG (closes #326)
- Fix all clippy warnings workspace-wide (zero warnings under -D warnings):
  explicit into_iter, too_many_arguments (#[allow]), derivable Default, collapsible
  match, map_or→is_some_and, unnecessary casts, vec_init_then_push, identity_op,
  is_multiple_of, div_ceil, repeat_n, manual_clamp, single_match, redundant_locals,
  needless_range_loop, default_constructed_unit_structs, enum_variant_names,
  manual_contains, let_and_return, needless_borrow, redundant_closure, print_literal,
  manual_pattern_char_comparison, unused imports/dead_code, doc_overindented_list_items
- Add #[allow(missing_docs)] on InstrKind and ConstantData (verbose variant fields)
- Add brief doc comments to all remaining undocumented public items
- Add //! # Examples no_run doc-tests to llvm-ir, llvm-transforms, llvm-bitcode crate roots
- Add hello_world.rs and opt_pipeline.rs examples to src/llvm/examples/
- Add Quick Start section to README.md
- Add Milestone L entry to CHANGELOG.md Unreleased section

+++ src/llvm-ir-parser/src/parser.rs
@@ -665,7 +665,7 @@ impl<'src> Parser<'src> {
                 }
             }
             Token::IntLit(n) => {
-                let n = *n as i64;
+                let n = *n;
                 let s = (n as u64).to_string();
                 self.lex.next()?;
                 let next = self.lex.next()?;
